Update to latest fuel-core and fuel-vm (#1670)
# Summary

Updates fuel-vm and fuel-core to latest versions. 

# Breaking Changes

Balances are now represented as u128.

# Checklist

- [x] All **changes** are **covered** by **tests** (or not applicable)
- [x] All **changes** are **documented** (or not applicable)
- [x] I **reviewed** the **entire PR** myself (preferably, on GH UI)
- [x] I **described** all **Breaking Changes** (or there's none)

---------

+++ e2e/tests/scripts.rs
@@ -140,7 +140,7 @@ async fn test_output_variable_estimation() -> Result<()> {
         .await?;
 
     let receiver_balance = receiver.get_asset_balance(&asset_id).await?;
-    assert_eq!(receiver_balance, amount);
+    assert_eq!(receiver_balance, amount as u128);
 
     Ok(())
 }
